+++ uizacoresdk/src/main/java/vn/uiza/uzv3/view/dlg/playlistfolder/AdapterPlaylistFolder.java
@@ -21,6 +21,7 @@ import java.util.List;
 import loitp.core.R;
 import vn.uiza.core.utilities.LAnimationUtil;
 import vn.uiza.core.utilities.LImageUtil;
+import vn.uiza.core.utilities.LLog;
 import vn.uiza.restapi.uiza.model.v3.metadata.getdetailofmetadata.Data;
 import vn.uiza.uzv3.util.UZUtil;
 
@@ -30,6 +31,8 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
     private int currentPositionOfDataList;
     private Context context;
     private CallbackPlaylistFolder callbackPlaylistFolder;
+    //private int sizeW;
+    //private int sizeH;
 
     public class PlayListHolder extends RecyclerView.ViewHolder {
         private TextView tvDuration;
@@ -59,6 +62,7 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
         this.dataList = dataList;
         this.currentPositionOfDataList = currentPositionOfDataList;
         this.callbackPlaylistFolder = callbackPlaylistFolder;
+        //sizeW = LScreenUtil.getScreenWidth() / 5;
     }
 
     @Override
@@ -69,10 +73,10 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
 
     @Override
     public void onBindViewHolder(final PlayListHolder playListHolder, final int position) {
+        //updateSizeFocus(playListHolder.rootView, false);
+
         final Data data = dataList.get(position);
         //LLog.d(TAG, "onBindViewHolder" + new Gson().toJson(item));
-
-        //playListHolder.tvDuration.setText(item.getDuration());
         UZUtil.setTextDuration(playListHolder.tvDuration, data.getDuration());
         //LLog.d(TAG, "item.getDuration(): " + item.getDuration());
         playListHolder.tvName.setText(data.getName());
@@ -93,20 +97,6 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
         } else {
             playListHolder.tvDescription.setText(data.getShortDescription());
         }
-
-        //RelativeLayout.LayoutParams rootLayoutParams = new RelativeLayout.LayoutParams((int) (sizeWRoot / 3.5), sizeHRoot);
-        //playListHolder.rootView.setLayoutParams(rootLayoutParams);
-
-        //RelativeLayout.LayoutParams layoutParams = new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, (int) (sizeWRoot / 3.5 / 2));
-        //playListHolder.ivCover.setLayoutParams(layoutParams);
-
-        /*String thumbnail;
-        if (data.getThumbnail() == null || data.getThumbnail().isEmpty()) {
-            thumbnail = Constants.URL_IMG_THUMBNAIL;
-        } else {
-            thumbnail = Constants.PREFIXS_SHORT + data.getThumbnail();
-        }*/
-        //LLog.d(TAG, "getThumbnail " + thumbnail);
         LImageUtil.load((Activity) context, data.getThumbnail(), playListHolder.ivCover);
 
         playListHolder.rootView.setOnClickListener(new View.OnClickListener() {
@@ -142,9 +132,9 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
         playListHolder.rootView.setOnFocusChangeListener(new View.OnFocusChangeListener() {
             @Override
             public void onFocusChange(View view, boolean isFocus) {
-                //LLog.d(TAG, "onFocusChange isFocus: " + isFocus);
+                LLog.d(TAG, "onFocusChange isFocus: " + isFocus);
+                //updateSizeFocus(playListHolder.rootView, isFocus);
                 if (isFocus) {
-                    LAnimationUtil.play(playListHolder.ivCover, Techniques.Pulse);
                     playListHolder.rootView.setBackgroundResource(R.drawable.bkg_item_playlist_folder);
                 } else {
                     playListHolder.rootView.setBackgroundResource(0);
@@ -160,4 +150,18 @@ public class AdapterPlaylistFolder extends RecyclerView.Adapter<AdapterPlaylistF
     public int getItemCount() {
         return dataList == null ? 0 : dataList.size();
     }
+
+    /*private void updateSizeFocus(LinearLayout linearLayout, boolean isFocus) {
+        if (linearLayout == null) {
+            return;
+        }
+        if (isFocus) {
+            linearLayout.getLayoutParams().width = sizeW + sizeW / 3;
+            //linearLayout.getLayoutParams().height = sizeH;
+        } else {
+            linearLayout.getLayoutParams().width = sizeW;
+            //linearLayout.getLayoutParams().height = sizeH / 2;
+        }
+        linearLayout.requestLayout();
+    }*/
 }