+++ uizacoresdk/src/main/java/vn/uiza/uzv3/view/rl/video/UZTVCallback.java
@@ -1,0 +1,7 @@
+package vn.uiza.uzv3.view.rl.video;
+
+import android.view.View;
+
+public interface UZTVCallback {
+    public void onFocusChange(View view, boolean isFocus);
+}