+++ uizacoresdk/src/main/java/vn/uiza/views/recyclerview/snappysmoothscroller/SnapType.java
@@ -1,0 +1,8 @@
+package vn.uiza.views.recyclerview.snappysmoothscroller;
+
+public enum SnapType {
+    START,
+    END,
+    CENTER,
+    VISIBLE
+}